+++ app/src/hooks/use-form-errors.ts
@@ -7,10 +7,18 @@ export type FormErrorsHandler<E> = ErrorHandler<E>;
 
 export type FormErrorsHandlers<E, T> = FieldErrorsHandler<E, T> | [FieldErrorsHandler<E, T>, FormErrorsHandler<E>];
 
-const useFormErrors = <E, T>(
+export type FieldsErrors<T> = { [F in keyof T]?: ReactNode };
+export type ClearFieldError<T> = (field: keyof T) => void;
+
+export type FormError = React.ReactNode;
+export type ClearFormError = () => void;
+
+type UseFormErrorsHook<E, T> = (
   handlers: FormErrorsHandlers<E, T>,
   error?: E,
-) => {
+) => [[FieldsErrors<T>, ClearFieldError<T>], [FormError, ClearFormError]];
+
+const useFormErrors = <E, T>(handlers: FormErrorsHandlers<E, T>, error?: E): ReturnType<UseFormErrorsHook<E, T>> => {
   const [fieldErrors, setFieldErrors] = useState<{ [F in keyof T]?: ReactNode }>({});
   const [formError, setFormError] = useState<ReactNode>();
 
@@ -21,9 +29,9 @@ const useFormErrors = <E, T>(
 
     const [fieldsHandler, formHandler] = Array.isArray(handlers) ? handlers : [handlers];
     const fields = Object.keys(fieldsHandler);
-    const fieldErrors: { [F in keyof T]?: ReactNode } = {};
+    const fieldErrors: FieldsErrors<T> = {};
 
-    fields.forEach((field) => {
+    fields.forEach(field => {
       const fieldError = fieldsHandler[field as keyof T](error);
 
       if (typeof fieldError !== 'undefined') {
@@ -45,7 +53,7 @@ const useFormErrors = <E, T>(
   return [
     [fieldErrors, (field: keyof T) => setFieldErrors({ ...fieldErrors, [field]: undefined })],
     [formError, () => setFormError(undefined)],
-  ] as const;
+  ];
 };
 
 export default useFormErrors;